HIVE-25717: INSERT INTO on external MariaDB/MySQL table fails silently
